Handle denied access where role insufficient

+++ src/Controller/Admin/CategoryController.php
@@ -4,16 +4,17 @@ namespace App\Controller\Admin;
 
 use App\Entity\Category;
 use App\Form\CategoryType;
+use Doctrine\DBAL\Exception;
 use App\Repository\CategoryRepository;
-use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
+use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 
 #[Route('/admin/category', name: 'admin_category_')]
 class CategoryController extends AbstractController
 {
-    #[Route("/list" , name: 'list')]
+    #[Route("/list", name: 'list')]
     public function categoryList(CategoryRepository $categoryRepository): Response
     {
         return $this->render('admin/category/index.html.twig', [
@@ -46,7 +47,7 @@ class CategoryController extends AbstractController
     {
         $form = $this->createForm(CategoryType::class, $category);
         $form->handleRequest($request);
-        if($form->isSubmitted() && $form->isValid()) {
+        if ($form->isSubmitted() && $form->isValid()) {
             $em = $this->getDoctrine()->getManager();
             $em->persist($category);
             $em->flush();
@@ -64,12 +65,25 @@ class CategoryController extends AbstractController
     #[Route('/delete/{id<\d+>}', name: 'delete')]
     public function deleteCategory(Category $category): Response
     {
-        $em = $this->getDoctrine()->getManager();
-        $em->remove($category);
-        $em->flush();
-
-        $this->addFlash('success', 'La catégorie à bien été supprimée');
-
+        if (!$this->isGranted('ROLE_SUPER_ADMIN')) {
+            $this->addFlash('error', 'Vous ne disposez pas des droits pour supprimer une catégorie !');
+        } else {
+            try {
+                $em = $this->getDoctrine()->getManager();
+                $em->remove($category);
+                $em->flush();
+                $this->addFlash('success', 'La catégorie à bien été supprimée');
+            } catch (Exception $e) {
+                $code = $e->getPrevious()->getCode();
+                if ($code == "23000") {
+                    $this->addFlash('error', 'Impossible de supprimer la catégorie car elle est encore utilisée !');
+                } else {
+                    $_ENV == 'dev' ?
+                        $this->addFlash('error', "Une erreur est survenue:" + $e->getPrevious()->getMessage())
+                        : $this->addFlash('error', "Une erreur est survenue, code erreur:" + $code + ". Merci de bien vouloir contacter l'administrateur.");
+                }
+            }
+        }
         return $this->redirectToRoute('admin_category_list');
     }
 }
